Merge branch 'release-2.2'

* release-2.2: (29 commits)
  Bump version for 2.2.10-pl release
  Increase modTransportPackage version columns range to smallint
  [#10211] Fix parser state bug triggered by media sources
  Remove executable bit from files
  Fixed loading modResource derivatives based on objects allowListingInClassKeyDropdown setting, not using showInContextMenu for that
  Fix js error on MODx.Console.close()
  Minor bug fixes
  Updated Dutch lexicons for 2.2.10
  Update update.class.php
  italian lexicon for modx 2.2.10
  Update file.inc.php
  Update setting.inc.php
  Fixed bugs and updated language file
  French translation by @ackwa
  Updated and fixed some of the Brazilian Portuguese translation files.
  Update Japanese lexicon topics
  Updated encoding for Brazilian Portuguese lexicon files
  Updating the lexicon files for Brazilian Portuguese
  Add a README.md
  Update German lexicon topics
  ...

+++ core/model/modx/transport/sqlsrv/modtransportpackage.map.inc.php
@@ -121,7 +121,7 @@ $xpdo_meta_map['modTransportPackage']= array (
     ),
     'version_major' => 
     array (
-      'dbtype' => 'tinyint',
+      'dbtype' => 'smallint',
       'phptype' => 'integer',
       'null' => false,
       'default' => 0,
@@ -129,7 +129,7 @@ $xpdo_meta_map['modTransportPackage']= array (
     ),
     'version_minor' => 
     array (
-      'dbtype' => 'tinyint',
+      'dbtype' => 'smallint',
       'phptype' => 'integer',
       'null' => false,
       'default' => 0,
@@ -137,7 +137,7 @@ $xpdo_meta_map['modTransportPackage']= array (
     ),
     'version_patch' => 
     array (
-      'dbtype' => 'tinyint',
+      'dbtype' => 'smallint',
       'phptype' => 'integer',
       'null' => false,
       'default' => 0,
@@ -154,7 +154,7 @@ $xpdo_meta_map['modTransportPackage']= array (
     ),
     'release_index' => 
     array (
-      'dbtype' => 'tinyint',
+      'dbtype' => 'smallint',
       'phptype' => 'integer',
       'null' => false,
       'default' => 0,
